requirements.txt
matplotlib==3.7.1
numpy==1.24.1
pandas==1.5.3
scikit-learn==1.2.2
seaborn==0.12.2
tqdm==4.65.0
streamlit
plotly
dice-ml
google-genai
python-dotenv
